+++ src/main/java/com/fiap/hawk/repository/ProjectRepository.java
@@ -9,9 +9,9 @@ import java.util.List;
 import java.util.Optional;
 
 public interface ProjectRepository extends MongoRepository<ProjectDocument, String> {
-	Optional<ProjectDocument> findByPublicId(Long publicId);
-	Optional<ProjectDocument> findByIdeaId(Long ideaId);
-	boolean existsByIdeaId(Long ideaId);
+	Optional<ProjectDocument> findByPublicId(Integer publicId);
+	Optional<ProjectDocument> findByIdeaId(Integer ideaId);
+	boolean existsByIdeaId(Integer ideaId);
 	List<ProjectDocument> findAllByOrderByCreatedAtDesc();
 	long countByStatus(ProjectStatus status);
 	long countByStage(ProjectStage stage);
