+++ src/main/java/com/fiap/hawk/repository/UserRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 public interface UserRepository extends MongoRepository<UserDocument, String> {
 	Optional<UserDocument> findByEmailIgnoreCase(String email);
-	Optional<UserDocument> findByPublicId(Long publicId);
+	Optional<UserDocument> findByPublicId(Integer publicId);
 	boolean existsByEmailIgnoreCase(String email);
 	List<UserDocument> findByRoleAndActive(Role role, boolean active);
 	List<UserDocument> findByActiveTrue();
